refactor: 3중 판정 집계를 lib 으로 이관
권한 단계 판정(124)이 이 계산을 유일한 입력으로 쓴다. 계산이 commands 에 있으면
lib 에서 그것을 import 해야 해서 역방향 의존이 생긴다. 계산은 순수하므로 lib 으로
내리고 commands/stats.ts 는 re-export 만 남긴다.

판정 계약의 원본이 갈리지 않게 stats.ts 에 재정의를 남기지 않았고, 같은 커밋에서
AutonomyStats 에 optional rollingSelfReportedOnly 를 additive 로 넣었다(RFC 0066 §4.4
승급 차단 신호). 기존 테스트 60건은 무수정으로 통과한다. RFC 0066 §2.1
